Possible mistake in explanation of code
The explanation above the block of code in the "Working with the Current Style of the Project" section says that the code "changes only the strokeWidth and creates another path" when in reality it changes strokeWidth and fillColor as well. Might be confusing hence corrected it. I have added the <code></code> tags around "fillColor" keeping with the convention.